requirements.txt


python-dotenv == 1.2.2


# ─────────────────────────────────────────
# CORE — LangGraph & LangChain
# ─────────────────────────────────────────
langchain == 1.2.15
langgraph == 1.1.6
langsmith == 0.7.29
langchain-core == 1.2.28
langchain-community == 0.4.1
langchain-opendataloader-pdf == 2.0.0
langchain-experimental == 0.4.1
langchain-huggingface == 1.2.2
langchain-pymupdf4llm == 0.5.0


# ─────────────────────────────────────────
# Preprocessing file — pdf, hmtl, docx
# ─────────────────────────────────────────
pymupdf4llm == 1.27.2.2
PyMuPDF == 1.27.2.2
camelot-py == 1.0.9
pdfplumber == 0.11.9
ollama
rapidocr-onnxruntime == 1.4.4
pytesseract == 0.3.13


sentence-transformers == 5.4.1


# ─────────────────────────────────────────
# VECTOR STORE — Qdrant
# ─────────────────────────────────────────
qdrant-client == 1.9.0

# ─────────────────────────────────────────
# RETRIEVAL — Keyword search BM25
# ─────────────────────────────────────────
rank-bm25 == 0.2.2


flashrank == 0.2.10

# ─────────────────────────────────────────
# GRAPH SEARCH — Relations entre articles
# ─────────────────────────────────────────
networkx == 3.6.1

# ─────────────────────────────────────────
# CONFIG — Settings & variables d'env
# ─────────────────────────────────────────
pydantic == 2.12.5
pydantic-settings == 2.13.1
python-dotenv == 1.2.2
pyyaml == 6.0.3              # dev.yaml / prod.yaml
 
# ─────────────────────────────────────────
# LOGGING — Logs structurés JSON
# ─────────────────────────────────────────
structlog>=24.4.0
rich>=13.9.0                  # console colorée en dev
 
# ─────────────────────────────────────────
# LLM — OpenRouter (OpenAI-compatible)
# ─────────────────────────────────────────
openai >= 1.30.0

# ─────────────────────────────────────────
# INTERFACE — Streamlit
# ─────────────────────────────────────────
streamlit == 1.56.0

# ─────────────────────────────────────────
# API — FastAPI + serveur
# ─────────────────────────────────────────
fastapi == 0.135.3
uvicorn[standard] == 0.44.0
httpx == 0.28.1               # client HTTP async (tests API)
 
# ─────────────────────────────────────────
# ÉVALUATION — RAGAS + benchmark
# ─────────────────────────────────────────
ragas == 0.4.3
pandas == 3.0.2            # analyse des résultats
numpy>=1.26.0
 
# ─────────────────────────────────────────
# UTILITAIRES TEXTE
# ─────────────────────────────────────────
tiktoken>=0.8.0               # comptage tokens (contexte LLM)
regex>=2024.9.11              # nettoyage textes juridiques
unidecode>=1.3.8              # normalisation accents
 
# ─────────────────────────────────────────
# TESTS
# ─────────────────────────────────────────
pytest == 9.0.3
pytest-asyncio == 1.3.0       # tests async (FastAPI, LangGraph)
pytest-cov == 7.1.0           # couverture de code
